Changed some of the logic in score.

+++ src/application/Main.java
@@ -178,7 +178,7 @@ public class Main extends Application {
 			} else if (gameEngine.getFoundPairs() == (gameEngine.getCards().length / 2)) {
 				gameEngine.stopTime();
 				time.setText(Long.toString(gameEngine.timePlayed())+" sec");
-				pointresult.setText(Long.toString(gameEngine.calculateScore()));
+				pointresult.setText(Double.toString(gameEngine.calculateScore()));
 			}
 
 		});
